+++ Proyecto Unity 3D/Assets/Scripts/Player/PlayerCollisions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10012ac96fd95c34282437b794690ef5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
